빌드 리포트 툴 업데이트 , 권한 설정 수정
빌드 리포트 툴 업데이트 , 권한 설정 수정

+++ MNP-QROAD/Assets/02. Scripts/System/Interface/ScissorCtrl.cs
@@ -36,7 +36,7 @@ public class ScissorCtrl : MonoBehaviour {
 
         //Application.targetFrameRate = 60;
 
-        float currentAspectRatio = (float)Screen.width / Screen.height;
+        float currentAspectRatio = (float)Screen.width / (float)Screen.height;
 
         //Debug.Log("★★★ ScissorCtrl currentAspectRatio :: " + currentAspectRatio);
